feat: add group talk functionality
Add basic group talk functionality; some features are yet to be
implemented.

Add the functionality to shut down
a TaskNode in the kernel, and introduce the 'at_terminated'
callback. Optimized the commander attribute in the kernel, allowing
for the direct addition of handler to the commander.

Allow each chat tab to have independent output.