fix(kernel)!: change the task_node parameter of the callback function to a keyword argument
When executing the callback function, if a task_node needs to be
inserted, insert it in the form of a keyword argument with the parameter
name "task_node". The original task_node was passed as the first
positional argument, but the first positional argument of a function may
have a special purpose. Therefore, it is better to change it to a
keyword argument. This also offers a more flexible and safer function
invocation style.