fix: fix group talk close bug and task_node remove bug
Take into account the situation where '_children' is empty in
'TaskNode.del_child'.

Fix 'the future belongs to a different loop than the one specified
as the loop argument' when close group talk. This is because
'GroupTalkManager' is used in two different conversations(event loops).

Fix the issue where the commander display does not update its status
when group talk is closed.