Fix group talk (#13)
* refactor: refactor the bead section

* fix: fix stream content in buffer was not clear at the end, fix the bead
not being deep copied

* fix(kernel)!: change the task_node parameter of the callback function to
a keyword argument

When executing the callback function, if a task_node needs to be
inserted, insert it in the form of a keyword argument with the parameter
name "task_node". The original task_node was passed as the first
positional argument, but the first positional argument of a function may
have a special purpose. Therefore, it is better to change it to a
keyword argument. This also offers a more flexible and safer function
invocation style.

* feat: add reset_role_prompt in GroupTalk plugin

* feat: make each 'chat_tab''s 'commander_display' display independently
fix: fix the issue where 'horse' cannot be closed when the size of
'apple' is greater than the size of 'happy'
fix: enable the delete conversation feature to handle cases where the
conversation is a group talk

* feat: add RotatingCube apple

* feat: add 'PENDING' and 'COMPLETED' states to TaskNode's state attribute
perf: optimize prompt for GroupTalk plugin
fix: fix bug where GroupTalk cannot be resumed if closed unexpectedly
style: change TaskNode's state attribute to uppercase

* fix: fix group talk close bug and task_node remove bug

Take into account the situation where '_children' is empty in
'TaskNode.del_child'.

Fix 'the future belongs to a different loop than the one specified
as the loop argument' when close group talk. This is because
'GroupTalkManager' is used in two different conversations(event loops).

Fix the issue where the commander display does not update its status
when group talk is closed.